Addressed concerns raised in timgraham's review
* Also includes some slight refactoring
