+++ tests/m2m_through_regress/tests.py
@@ -6,7 +6,7 @@ from django.test import TestCase
 from django.utils.six import StringIO
 
 from .models import (
-    Car, CarDriver, CompetingTeam, Driver, Event, Group, IndividualCompetitor, Membership, Person, UserMembership,
+    Car, CarDriver, Driver, Group, Membership, Person, UserMembership,
 )
 
 
@@ -271,47 +271,3 @@ class ThroughLoadDataTestCase(TestCase):
             ': 1}}, {"pk": 1, "model": "m2m_through_regress.person", "fields": {"name": "Guido"}}, {"pk": 1, '
             '"model": "m2m_through_regress.group", "fields": {"name": "Python Core Group"}}]'
         )
-
-
-class M2MThroughMultiTableTestCase(TestCase):
-    @classmethod
-    def setUpTestData(cls):
-        cls.alice = Person.objects.create(name='Alice')
-        cls.bob = Person.objects.create(name='Bob')
-        cls.chris = Person.objects.create(name='Chris')
-        cls.team_alpha = Group.objects.create(name='Alpha')
-        Membership.objects.create(person=cls.alice, group=cls.team_alpha)
-        Membership.objects.create(person=cls.bob, group=cls.team_alpha)
-        cls.event = Event.objects.create(name='Exposition Match')
-        IndividualCompetitor.objects.create(event=cls.event, person=cls.chris, disqualified=False)
-        CompetingTeam.objects.create(event=cls.event, team=cls.team_alpha, disqualified=True)
-
-    def test_m2m_query(self):
-        result = self.event.teams.all()
-        self.assertEqual(len(result), 1)
-        self.assertEqual(result[0], self.team_alpha)
-
-    def test_m2m_reverse_query(self):
-        result = self.chris.event_set.all()
-        self.assertEqual(len(result), 1)
-        self.assertEqual(result[0], self.event)
-
-    def test_m2m_query_proxied(self):
-        result = self.event.special_people.all()
-        self.assertEqual(len(result), 1)
-        self.assertEqual(result[0], self.chris)
-
-    def test_m2m_reverse_query_proxied(self):
-        result = self.chris.special_event_set.all()
-        self.assertEqual(len(result), 1)
-        self.assertEqual(result[0], self.event)
-
-    def test_m2m_prefetch_proxied(self):
-        result = Event.objects.filter(name='Exposition Match').prefetch_related('special_people')
-        # Trigger Queryset evaluation
-        str(result)
-
-    def test_m2m_prefetch_reverse_proxied(self):
-        result = Person.objects.filter(name='Alice').prefetch_related('special_event_set')
-        # Trigger Queryset evaluation
-        str(result)
